refactor HourRow to own component
